Renamed input_partition submodule to just simply be inputs
