+++ docs/source/api/models/litter/inputs.md
