+++ virtual_ecosystem/models/litter/__init__.py
@@ -11,8 +11,8 @@ Each of the litter sub-modules has its own API reference page:
   pools that the litter model is comprised of.
 * :mod:`~virtual_ecosystem.models.litter.chemistry` tracks the chemistry (lignin,
   nitrogen and phosphorus) of the litter pools.
-* :mod:`~virtual_ecosystem.models.litter.input_partition` handles the partitioning
-  of biomass input between the different litter pools.
+* :mod:`~virtual_ecosystem.models.litter.inputs` handles the partitioning of biomass
+  input between the different litter pools.
 * :mod:`~virtual_ecosystem.models.litter.env_factors` provides the functions
   capturing the impact of environmental factors on litter decay.
 * :mod:`~virtual_ecosystem.models.litter.constants` provides a set of dataclasses
